Fix all compiler warnings + clang-format entire codebase
Warning fixes (3506 -> 0):
- Remove const from value return types (DasherTypes.h, AlphInfo.h)
- Add virtual destructors to 5 base classes (Messages, FrameRate, AlphabetMap, Trainer)
- Fix sign-compare in AlphIO, GameModule, MandarinAlphMgr, ScreenGameModule
- Fix char-subscripts in AlphabetMap (cast to unsigned char)
- Fix shadow variables in AlphabetManager, ConversionManager, PPMLanguageModel,
  PPMPYLanguageModel, Trainer, UserLogTrial
- Fix overloaded-virtual in AlternatingDirectMode, CTWLanguageModel
- Fix reorder-ctor in 9 constructors across 8 files
- Remove dead code (unused vars, unused functions)
- Add -Wno-unused-parameter (standard for virtual override signatures)

Also:
- clang-format -i on all src/ and tests/ files
- Add build-sanitize/ and Testing/ to .gitignore
- Add -Wno-unused-parameter to CMakeLists.txt

+++ tests/test_color_math.cpp
@@ -31,7 +31,7 @@ TEST(color_alpha_zero) {
 
 TEST(color_black) {
     int c = dasher_color_rgb(0, 0, 0);
-    ASSERT_EQ(c, 0xFF000000);
+    ASSERT_EQ(c, (int)0xFF000000);
     ASSERT_EQ(dasher_color_get_alpha(c), 255);
     ASSERT_EQ(dasher_color_get_red(c), 0);
     ASSERT_EQ(dasher_color_get_green(c), 0);
@@ -41,7 +41,7 @@ TEST(color_black) {
 
 TEST(color_white) {
     int c = dasher_color_rgb(255, 255, 255);
-    ASSERT_EQ(c, 0xFFFFFFFF);
+    ASSERT_EQ(c, (int)0xFFFFFFFF);
     ASSERT_EQ(dasher_color_get_alpha(c), 255);
     ASSERT_EQ(dasher_color_get_red(c), 255);
     ASSERT_EQ(dasher_color_get_green(c), 255);
